requirements.txt
# PageIndex RAG
pageindex

# Traditional RAG
chromadb
sentence-transformers
pypdf2

# LLM
groq

# Frontend
streamlit
markdown

# Utilities
python-dotenv
requests
reportlab
